Moved UI helpers to the base since the same helpers will be used in MediaType
